refactor: move use-scroll-lock to modal
